+++ src/main/java/servlets/StudentEnrollmentServlet.java
@@ -5,6 +5,8 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import models.User;
+import models.UserType;
 
 import java.io.IOException;
 
@@ -19,10 +21,16 @@ public class StudentEnrollmentServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        if (request.getParameter("enrollment").equals("enroll"))
-            mySQLEnrollmentDao.save(request.getParameter("courseId"), request.getParameter("studentId"));
-        else if (request.getParameter("enrollment").equals("unEnroll"))
-            mySQLEnrollmentDao.delete(request.getParameter("courseId"), request.getParameter("studentId"));
-        response.sendRedirect("/course/" + request.getParameter("courseId"));
+        User user = (User) request.getSession().getAttribute("user");
+
+        if (user != null && user.getUserType() == UserType.INSTRUCTOR) {
+            if (request.getParameter("enrollment").equals("enroll"))
+                mySQLEnrollmentDao.save(request.getParameter("courseId"), request.getParameter("studentId"));
+            else if (request.getParameter("enrollment").equals("unEnroll"))
+                mySQLEnrollmentDao.delete(request.getParameter("courseId"), request.getParameter("studentId"));
+            response.sendRedirect("/course/" + request.getParameter("courseId"));
+        }
+        else if (user == null) response.sendRedirect("/login");
+        else response.sendRedirect("/errorHandler?errorCode=403&errorMessage=Not authorized");
     }
 }
